FIX: Refactored parser test into several files for testing features individually
