+++ tests/unit/parser/parser-card-mix.test.ts
@@ -1,0 +1,31 @@
+import { CardType } from "src/card/questions/question";
+
+import { parserOptions, parseT } from "tests/unit/helpers/unit-test-parser-helper";
+
+test("Test parsing of a mix of card types", () => {
+    expect(
+        parseT(
+            "# Lorem Ipsum\n\nLorem ipsum dolor ==sit amet==, consectetur ==adipiscing== elit.\n" +
+            "Duis magna arcu, eleifend rhoncus ==euismod non,==\nlaoreet vitae enim.\n\n" +
+            "Fusce placerat::velit in pharetra gravida\n\n" +
+            "Donec dapibus ullamcorper aliquam.\n??\nDonec dapibus ullamcorper aliquam.\n<!--SR:2021-08-11,4,270-->",
+            parserOptions,
+        ),
+    ).toEqual([
+        [
+            CardType.Cloze,
+            "Lorem ipsum dolor ==sit amet==, consectetur ==adipiscing== elit.\n" +
+            "Duis magna arcu, eleifend rhoncus ==euismod non,==\n" +
+            "laoreet vitae enim.",
+            2,
+            4,
+        ],
+        [CardType.SingleLineBasic, "Fusce placerat::velit in pharetra gravida", 6, 6],
+        [
+            CardType.MultiLineReversed,
+            "Donec dapibus ullamcorper aliquam.\n??\nDonec dapibus ullamcorper aliquam.\n<!--SR:2021-08-11,4,270-->",
+            8,
+            11,
+        ],
+    ]);
+});
