+++ tests/unit/parser/parser-debugger.test.ts
@@ -1,0 +1,29 @@
+import { QuestionParser, setDebugParser } from "src/parser";
+
+import { parserOptions } from "tests/unit/helpers/unit-test-parser-helper";
+
+describe("Parser debug messages", () => {
+    test("Messages disabled", () => {
+        // replace console error log with an empty mock function
+        const logSpy = jest.spyOn(global.console, "log").mockImplementation(() => { });
+        setDebugParser(false);
+
+        QuestionParser.parse("", parserOptions);
+        expect(logSpy).toHaveBeenCalledTimes(0);
+
+        // restore original console error log
+        logSpy.mockRestore();
+    });
+
+    test("Messages enabled", () => {
+        // replace console error log with an empty mock function
+        const logSpy = jest.spyOn(global.console, "log").mockImplementation(() => { });
+        setDebugParser(true);
+
+        QuestionParser.parse("", parserOptions);
+        expect(logSpy).toHaveBeenCalled();
+
+        // restore original console error log
+        logSpy.mockRestore();
+    });
+});
